+++ authority-portal-frontend/src/app/pages/dashboard-page/dashboard-page.module.ts
@@ -1,6 +1,7 @@
 import {CommonModule} from '@angular/common';
 import {NgModule} from '@angular/core';
 import {PortalLayoutModule} from 'src/app/common/layouts/portal-layout/portal-layout.module';
+import {SharedModule} from 'src/app/shared/shared.module';
 import {ImageSliderModule} from '../../common/components/image-slider/image-slider.module';
 import {DashboardPageComponent} from './dashboard-page/dashboard-page.component';
 
@@ -12,6 +13,7 @@ import {DashboardPageComponent} from './dashboard-page/dashboard-page.component'
     // Authority Portal
     ImageSliderModule,
     PortalLayoutModule,
+    SharedModule,
   ],
   declarations: [DashboardPageComponent],
   exports: [DashboardPageComponent],