+++ authority-portal-frontend/src/app/pages/dashboard-page/dashboard-page/dashboard-page.component.ts
@@ -1,31 +1,25 @@
-import {Component} from '@angular/core';
-import {SlideInterface} from '../../../common/components/image-slider/image-slider/slide.interface';
+import {Component, Inject} from '@angular/core';
+import {DomSanitizer} from '@angular/platform-browser';
+import {APP_CONFIG, AppConfig} from 'src/app/core/config/app-config';
+import {IframeConfig} from 'src/app/shared/components/common/iframe/iframe.model';
 
 @Component({
   selector: 'app-dashboard',
   templateUrl: './dashboard-page.component.html',
 })
 export class DashboardPageComponent {
-  slides: SlideInterface[] = [
-    {
-      url: 'https://images.pexels.com/photos/5244314/pexels-photo-5244314.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1',
-      title: 'View and manage all users assigned to the company.',
-    },
-    {
-      url: 'https://images.pexels.com/photos/7580715/pexels-photo-7580715.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1',
-      title: 'Get to know the supported use cases to start your journey',
-    },
-    {
-      url: 'https://images.pexels.com/photos/8386422/pexels-photo-8386422.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1',
-      title: 'Transparent services and applications for your ideas',
-    },
-    {
-      url: 'https://images.pexels.com/photos/10156134/pexels-photo-10156134.jpeg?auto=compress&cs=tinysrgb&w=1600',
-      title: 'Join our large ecosystem of modern applications',
-    },
-    {
-      url: 'https://images.pexels.com/photos/532173/pexels-photo-532173.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1',
-      title: 'Control your data in transparent and intuitive way',
-    },
-  ];
+  iframeConfigData: IframeConfig;
+
+  constructor(
+    @Inject(APP_CONFIG) public appConfig: AppConfig,
+    private sanitizer: DomSanitizer,
+  ) {
+    this.iframeConfigData = {
+      srcUrl: this.sanitizer.bypassSecurityTrustResourceUrl(
+        this.appConfig.iframeUrl,
+      ),
+      width: '100%',
+      height: '100%',
+    };
+  }
 }