+++ authority-portal-frontend/src/app/shared/components/common/iframe/iframe.model.ts
@@ -1,0 +1,7 @@
+import {SafeResourceUrl} from '@angular/platform-browser';
+
+export interface IframeConfig {
+  srcUrl: SafeResourceUrl;
+  width: string;
+  height: string;
+}